format price to %.2f in b5mo and b5mp SCDs, not in merged and raw ones.

+++ b5m/start_b5m_matcher.rb
@@ -7,6 +7,7 @@ require_relative 'b5m_helper'
 
 
 top_dir = File.dirname(File.expand_path(__FILE__))
+stop_script = File.join(top_dir, "stop_b5m_matcher.rb")
 log_dir = File.join(top_dir, "log")
 Dir.mkdir(log_dir) unless File.exists?(log_dir)
 pid_file = File.join(log_dir, "pid")
@@ -16,28 +17,47 @@ end
 at_exit do
   FileUtils.rm_rf(pid_file)
 end
-config = nil
-config_file = File.join(top_dir, "config.yml")
-default_config_file = File.join(top_dir, "config.yml.default")
-if File.exist?(config_file)
-  config = YAML::load(File.open(config_file))
-elsif File.exist?(default_config_file)
-  config = YAML::load(File.open(default_config_file))
-end
-if config.nil?
-  puts 'config file not found'
-  exit(false)
-end
 
+#Signal.trap("INT") do
+  #cmd = "ruby #{stop_scripts}"
+  #puts "calling #{cmd}"
+  #system(cmd)
+#end
+
+config = nil
+config_file = nil
 doreindex = false
 domerge = true
-ARGV.each do |a|
+dob5mc = true
+dologserver = true
+ARGV.each_with_index do |a,i|
   if a=="--reindex"
     doreindex = true
   elsif a=="--nomerge"
     domerge = false
+  elsif a=="--nob5mc"
+    dob5mc = false
+  elsif a=="--nologserver"
+    dologserver = false
+  elsif a=="--config"
+    config_file = ARGV[i+1]
   end
 end
+
+if config_file.nil?
+  config_file = File.join(top_dir, "config.yml")
+  unless File.exists?(config_file)
+    config_file = File.join(top_dir, "config.yml.default")
+  end
+end
+config_file = nil unless File.exists?(config_file)
+if config_file.nil?
+  abort 'config file not found'
+else
+  puts "Use config #{config_file}"
+  config = YAML::load_file(config_file)
+end
+
 
 match_config = config["matcher"]
 log_server = match_config["log_server"]
@@ -71,7 +91,7 @@ matcher_program = File.join(top_dir, "b5m_matcher")
 merger_program = File.join(File.dirname(top_dir), "scripts", "ScdMerger")
 FileUtils.rm_rf("#{match_path['b5m_scd']}")
 FileUtils.rm_rf(work_dir) if doreindex
-Dir.mkdir(work_dir) unless File.exist?(work_dir)
+FileUtils.mkdir_p(work_dir) unless File.exist?(work_dir)
 Dir.mkdir(db_path) unless File.exist?(db_path)
 Dir.mkdir(mdb) unless File.exist?(mdb)
 mdb_instance_list = []
@@ -90,7 +110,7 @@ merged_scd = File.join(mdb_instance, "merged")
 raw_scd = File.join(mdb_instance, "raw")
 comment_merged_scd = File.join(mdb_instance, "comment_merged")
 if File.exist?(mdb_instance)
-  exit(-1)
+  abort "#{mdb_instance} already exists"
 else
   Dir.mkdir(mdb_instance)
   Dir.mkdir(merged_scd)
@@ -157,8 +177,10 @@ if domerge
   if reindex
     puts "reindex!!"
     system("#{merger_program} #{scd} #{ENV['B5MO_PROPERTY']} #{merged_scd}")
+    abort("merge scd failed") unless $?.success?
   else
     system("#{merger_program} #{scd} #{ENV['B5MO_PROPERTY']} #{merged_scd} --gen-all")
+    abort("merge scd failed") unless $?.success?
   end
 else
   system("cp #{scd}/*.SCD #{merged_scd}/")
@@ -166,10 +188,13 @@ end
 
 #generate raw scds
 system("#{matcher_program} --raw-generate -S #{merged_scd} --raw #{raw_scd} --odb #{odb}")
+abort("generating raw scd failed") unless $?.success?
 
 #split SCDs
 system("#{matcher_program} -P -N T -S #{train_scd} -K #{work_dir}")
+abort("spliting T scd failed") unless $?.success?
 system("#{matcher_program} -P -N A -S #{raw_scd} -K #{work_dir}")
+abort("spliting A scd failed") unless $?.success?
 
 task_list.each do |task|
   next unless task.valid
@@ -190,16 +215,20 @@ task_list.each do |task|
     ofs.puts(task.info['name'])
     ofs.close
     system("#{matcher_program} -M -S #{raw_scd} -K #{category_dir}")
+    abort("complete matching failed") unless $?.success?
   elsif task.type == CategoryTask::SIM
     #do similarity matching
     puts "start similarity matching #{task.cid}"
     system("#{matcher_program} -I -C #{cma} -S #{raw_scd} -K #{category_dir}")
+    abort("similarity matching failed") unless $?.success?
   else
     next unless task.info['disable'].nil?
     puts "start building attribute index for #{task.cid}"
     system("#{matcher_program} -A -Y #{synonym} -C #{cma} -S #{train_scd} -K #{category_dir}")
+    abort("attribute indexing failed") unless $?.success?
     puts "start matching for #{task.cid}"
     system("#{matcher_program} -B -Y #{synonym} -C #{cma} -S #{raw_scd} -K #{category_dir}")
+    abort("attribute matching failed") unless $?.success?
   end
 
 end
@@ -217,35 +246,42 @@ Dir.mkdir(b5mc_scd_instance)
 
 #b5mo generator
 system("#{matcher_program} --b5mo-generate -S #{raw_scd} --b5mo #{b5mo_scd_instance} -K #{mdb_instance}")
+abort("b5mo generate failed") unless $?.success?
+system("rm -rf #{b5mo_scd}/*")
+system("cp #{b5mo_scd_instance}/* #{b5mo_scd}/")
 
 #uue generator
 system("#{matcher_program} --uue-generate --b5mo #{b5mo_scd_instance} --uue #{mdb_instance}/uue --odb #{odb}")
+abort("uue generate failed") unless $?.success?
 
 #b5mp generator, update odb and pdb here.
 system("#{matcher_program} --b5mp-generate --b5mo #{b5mo_scd_instance} --b5mp #{b5mp_scd_instance} --uue #{mdb_instance}/uue --odb #{odb} --pdb #{pdb}")
+system("rm -rf #{b5mp_scd}/*")
+system("cp #{b5mp_scd_instance}/* #{b5mp_scd}/")
 odb_in_mdb = File.join(mdb_instance, "odb")
 pdb_in_mdb = File.join(mdb_instance, "pdb")
 FileUtils.cp_r(odb, odb_in_mdb)
 FileUtils.cp_r(pdb, pdb_in_mdb)
 
 #merge comment scd
-if domerge
-  system("#{merger_program} #{comment_scd} #{ENV['B5MC_PROPERTY']} #{comment_merged_scd}")
-else
-  system("cp #{comment_scd}/*.SCD #{comment_merged_scd}/")
+if dob5mc
+  if domerge
+    system("#{merger_program} #{comment_scd} #{ENV['B5MC_PROPERTY']} #{comment_merged_scd}")
+    abort("merge comment scd failed") unless $?.success?
+  else
+    system("cp #{comment_scd}/*.SCD #{comment_merged_scd}/")
+  end
+
+  #b5mc generator
+  system("#{matcher_program} --b5mc-generate --b5mc #{b5mc_scd_instance} -S #{comment_merged_scd} --odb #{odb}")
+  abort("b5mc generate failed") unless $?.success?
+  system("rm -rf #{b5mc_scd}/*")
+  system("cp #{b5mc_scd_instance}/* #{b5mc_scd}/")
 end
 
-#b5mc generator
-system("#{matcher_program} --b5mc-generate --b5mc #{b5mc_scd_instance} -S #{comment_merged_scd} --odb #{odb}")
-
-#copy generated SCDs
-system("rm -rf #{b5mo_scd}/*")
-system("rm -rf #{b5mp_scd}/*")
-system("rm -rf #{b5mc_scd}/*")
-system("cp #{b5mo_scd_instance}/* #{b5mo_scd}/")
-system("cp #{b5mp_scd_instance}/* #{b5mp_scd}/")
-system("cp #{b5mc_scd_instance}/* #{b5mc_scd}/")
-
-#logserver update
-system("#{matcher_program} --logserver-update --logserver-config '#{log_server}' --uue #{mdb_instance}/uue --odb #{odb}")
+if dologserver
+  #logserver update
+  system("#{matcher_program} --logserver-update --logserver-config '#{log_server}' --uue #{mdb_instance}/uue --odb #{odb}")
+  abort("log server update failed") unless $?.success?
+end
